refactor(commands): split slash command handlers
- extract command handlers into focused core/config/io/mcp modules
- move MCP prompt parsing and validation into a dedicated parser module
- centralize handler utilities for usage status and arg requirements
- rewire command registry handlers to the new module layout
- update command tests and exports for moved command functions
